requirements.txt
antlr4-python3-runtime==4.13.0
pytest
streamlit
plotly